requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.5.0
uritemplate==4.1.1
whitenoise==6.6.0
